[CCFPCM-0340] alerts lambda (WIP)
