+++ apps/backend/src/notification/entities/program-required-file.entity.ts
@@ -1,4 +1,3 @@
-import { FileTypes } from '../../constants';
 import {
   Column,
   CreateDateColumn,
@@ -9,6 +8,7 @@ import {
   Relation,
 } from 'typeorm';
 import { FileIngestionRulesEntity } from './file-ingestion-rules.entity';
+import { FileTypes } from '../../constants';
 
 @Entity('program_required_file')
 export class ProgramRequiredFileEntity {
